apps/api/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
astral==3.2
asttokens==2.4.1
astunparse==1.6.3
asyncpg==0.29.0
bcrypt==4.2.0
boto3==1.40.74
botocore==1.40.74
cachetools==6.2.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.3.0
cloudpickle==3.1.0
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
dnspython==2.8.0
email_validator==2.1.1
executing==2.0.1
fastapi==0.115.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.47.2
fsspec==2024.12.0
gast==0.5.4
google-api-core==2.28.1
google-auth==2.42.1
google-cloud==0.34.0
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-core==2.5.0
google-cloud-logging==3.12.1
google-cloud-secret-manager==2.21.1
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.11.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.7
imageio==2.37.0
imbalanced-learn==0.13.0
importlib_metadata==8.7.0
ipykernel==6.29.2
ipython==8.21.0
jedi==0.19.1
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==3.3.3
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
lime==0.2.0.1
llvmlite==0.43.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.3
opencv-python==4.9.0.80
opentelemetry-api==1.38.0
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.8
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg-pool==3.2.7
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.17.2
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.3
pyzmq==25.1.2
requests==2.32.3
rich==13.7.1
rsa==4.9.1
s3transfer==0.14.0
scikit-image==0.25.2
scikit-learn==1.4.0
scipy==1.11.4
seaborn==0.13.2
setuptools==70.0.0
shap==0.47.2
six==1.16.0
sklearn-compat==0.1.3
slicer==0.0.8
sniffio==1.3.1
stack-data==0.6.3
starlette==0.38.6
sympy==1.13.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2025.2.18
torch==2.5.1
tornado==6.4
tqdm==4.67.1
traitlets==5.14.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.4
urllib3==2.2.1
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
wxPython==4.2.1
zipp==3.23.0

# PDF generation
reportlab==4.0.7

# Excel generation
openpyxl==3.1.2

# Document generation
python-docx==1.1.0

# Google Cloud Pub/Sub (for event bus)
google-cloud-pubsub==2.18.4

# Google Cloud Tasks (for task queue)
google-cloud-tasks==2.14.2

# Pydantic Settings
pydantic-settings==2.1.0

# JWT Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Email
sendgrid==6.11.0

# SQLAlchemy
sqlalchemy==2.0.23
alembic==1.13.1
